gui: move to gui submodule, split simulation into it's own file
