+++ pyproject.toml
@@ -16,6 +16,7 @@ dependencies = [
     "galois >= 0.3.0",
     "numpy >= 1.21.0",
     "pyqt5",
+    "attrs >= 20.0.0"
 ]
 classifiers = [
     "Programming Language :: Python :: 3",
